cw_vm -> cycles

+++ include/corewar.h
@@ -6,7 +6,7 @@
 /*   By: nfinkel <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/13 16:16:50 by nfinkel           #+#    #+#             */
-/*   Updated: 2018/03/19 16:56:17 by lfabbro          ###   ########.fr       */
+/*   Updated: 2018/03/19 17:27:55 by lfabbro          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ typedef struct		s_cw
 	t_proc			*prev;
 	t_proc			*current;
 	t_proc			*procs;
-	size_t			cycle;
+	int				cycle;
 	int				cycle_to_die;
 	t_opt			opt;
 	uint8_t			n_champs;
